Extract product config CLI commands
